fix: network tests
